Review symlinks

+++ user-install/install-java.sh
@@ -56,8 +56,10 @@ function install_java_binaries {
   if [ ! -d ${tools}/${folder} ] ;then
     tar -C ${tools} -xpf ${archive}
   fi
-  if [ -L ${tools}/${symlink} ] ;then rm ${tools}/${symlink} ;fi
-  ln -s ${folder} ${tools}/${symlink}
+  if [ ! -z ${symlink} ] ;then
+    if [ -L ${tools}/${symlink} ] ;then rm ${tools}/${symlink} ;fi
+    ln -s ${folder} ${tools}/${symlink}
+  fi
 }
 
 function install_java {
